+++ Libftprintf/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: lwiller <lwiller@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/04 06:59:10 by lwiller           #+#    #+#             */
-/*   Updated: 2021/01/05 13:14:18 by lwiller          ###   ########lyon.fr   */
+/*   Updated: 2021/01/08 10:41:34 by lwiller          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int		ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 int		ft_isdigit(int c);
-char	*ft_itoa(int n);
+char	*ft_itoa(long int n);
 int 	ft_putchar_int(char c);
 int		ft_putstr_int(char *str);
 char	*ft_strdup(char *str);
